+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Tapping on a tab bar icon can let you scroll to top.
+
 ## [0.1 (95)] - 2023-11-27Z
 
 - Fixed a bug where a root note could be rendered as a reply